Rename the helper module to statusToTagColour
